// src/pages/InstituteHistoryPage/types.ts
export interface InstituteHistoryData {
  title: string
  firstParagraph: string
  secondParagraph: string
  thirdParagraph: string
  fourthParagraph: string
  fifthParagraph: string
  sixthParagraph: string
  seventhParagraph: string
  eighthParagraph: string
  ninthParagraph: string
  tenthParagraph: string
  eleventhParagraph: string
  twelfthParagraph: string
  thirteenthParagraph: string
  fourteenthParagraph: string
  fifteenthParagraph: string
  sixteenthParagraph: string
  seventeenthParagraph: string
  eighteenthParagraph: string
  nineteenthParagraph: string
  twentiethParagraph: string
  twentyFirstParagraph: string
  twentySecondParagraph: string
  twentyThirdParagraph: string
  twentyFourthParagraph: string
  twentyFifthParagraph: string
  twentySixthParagraph: string
  twentySeventhParagraph: string
  twentyEighthParagraph: string
  twentyNinthParagraph: string
  thirtiethParagraph: string
  imageCaption: string
  imageAlt: string
  secondImageCaption: string
  secondImageAlt: string
  thirdImageCaption: string
  thirdImageAlt: string
  fourthImageCaption: string
  fourthImageAlt: string
  fifthImageCaption: string
  fifthImageAlt: string
  sixthImageCaption: string
  sixthImageAlt: string
} 